requirements.txt
email-validator==1.1.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.0
requests==2.27.0
WTForms==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
